+++ chunking/utils/relay/types.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+
+class RelayMessage(BaseModel):
+    document_hash: str
+    embeddings: list[list[float]]
+
+class RelayPayload(BaseModel):
+    message: RelayMessage
+    signature: str
+
