+++ tests/test_relay_round.py
@@ -5,7 +5,7 @@ from re import A
 
 from openai import AsyncOpenAI, OpenAI
 from chunking.protocol import chunkSynapse
-from chunking.validator.relay import make_relay_payload
+from chunking.utils.relay.relay import make_relay_payload
 from chunking.validator.task_api import (
     Task,
     calculate_chunk_qty,
